Simplify the golang dependencies
This removes the dependency on kubernetes/kubernetes
repository and replaces it with:

- bundled cpuset
  from https://github.com/kubernetes/kubernetes/blob/759e043136b249f9e19355d3b2c9261b3ac82a70/pkg/kubelet/cm/cpuset/cpuset.go
- bundled pod resources api client
  from https://github.com/kubernetes/kubernetes/blob/52457842d155743f0e3fc57ade87251cca37d375/pkg/kubelet/apis/podresources/client.go#L56
- kubernetes/kubelet (APIs only)

The bundled code is placed into the pkg/k8s_imported directory to
make it obvious where it is coming from.

The effect is a significant reduction of indirect dependencies
allowing easier consumption by other projects (like must gather).

Signed-off-by: Martin Sivak <msivak@redhat.com>